requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.1
defusedxml==0.7.1
Django==4.1.2
django-allauth==0.47.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
djongo==1.3.6
dnspython==2.2.1
facebook-sdk==3.1.0
google-api-core==1.33.2
google-api-python-client==1.12.1
google-auth==2.13.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
idna==3.4
mongoengine==0.24.2
oauthlib==3.2.1
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.5.0
pymongo==3.12.1
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.2.4
types-cryptography==3.3.23.1
tzdata==2022.5
uritemplate==3.0.1
urllib3==1.26.12